+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- The stoppable_sleep method of the wizard tests does not block the main thread.
+
 ## [0.8.0] - 2024-04-23
 ### Added
 - PDOs for the EtherCAT protocol.