+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Set the boot_in_app flag according to the file extension.
+- FSoE module.
 
 ### Fix
 - Bug retrieving interface adapter name in Linux.